pyproject.toml
[project]
name = "cours-manticore"
version = "0.1.0"
description = "Opération Manticore — Cours bases de données non-relationnelles"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary",
    "geopandas",
    "pandas",
    "pyarrow",
    "neo4j",
    "python-dotenv",
    "ipykernel",
    "matplotlib",
    "duckdb>=1.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
dev-dependencies = [
    "jupyter",
    "notebook",
]
